pyproject.toml

[tool.poetry]
name = "text_includer"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.4.0"
openpyxl = "^3.1.2"
xlutils = "^2.0.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.21.2"
pre-commit = "^4.1.0"
black = "^25.1.0"
flake8 = "^7.1.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line_length = 120
